# Improve serializer performance