--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/NullableConverter.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/NullableConverter.cs
@@ -72,5 +72,6 @@ namespace System.Text.Json.Serialization.Converters
         }
 
         internal override bool IsNull(in T? value) => !value.HasValue;
+        internal override bool IsNullableOfT() => true;
     }
 }