--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/NullableConverter.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/NullableConverter.cs
@@ -70,8 +70,5 @@ namespace System.Text.Json.Serialization.Converters
                 _converter.WriteNumberWithCustomHandling(writer, value.Value, handling);
             }
         }
-
-        internal override bool IsNull(in T? value) => !value.HasValue;
-        internal override bool IsNullableOfT() => true;
     }
 }