--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterOfT.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterOfT.cs
@@ -3,6 +3,7 @@
 
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 using System.Text.Json.Serialization.Metadata;
 
 namespace System.Text.Json.Serialization
@@ -29,6 +30,12 @@ namespace System.Text.Json.Serialization
                 HandleNullOnRead = true;
                 HandleNullOnWrite = true;
             }
+            else
+            {
+                _autoWriteNulls = CanBeNull;
+            }
+
+            _isNullableOfT = IsNullableOfT();
 
             // For the HandleNull == false case, either:
             // 1) The default values are assigned in this type's virtual HandleNull property
@@ -38,6 +45,17 @@ namespace System.Text.Json.Serialization
 
             CanUseDirectReadOrWrite = !CanBePolymorphic && IsInternalConverter && ConverterStrategy == ConverterStrategy.Value;
         }
+
+        // <summary>
+        // Optimized check for 'CanBeNull && !HandleNullOnWrite'
+        // </summary>
+        private bool _autoWriteNulls;
+
+
+        // <summary>
+        // Optimized check to check if converter is for Nullable<T>
+        // </summary>
+        private bool _isNullableOfT;
 
         /// <summary>
         /// Determines whether the type can be converted.
@@ -143,6 +161,9 @@ namespace System.Text.Json.Serialization
         /// <remarks>Note that the value of <seealso cref="HandleNull"/> determines if the converter handles null JSON tokens.</remarks>
         public abstract T? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options);
 
+#if NET6_0_OR_GREATER
+        [MethodImpl(MethodImplOptions.AggressiveOptimization)]
+#endif
         internal bool TryRead(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options, ref ReadStack state, out T? value)
         {
             if (ConverterStrategy == ConverterStrategy.Value)
@@ -310,10 +331,20 @@ namespace System.Text.Json.Serialization
         }
 
         /// <summary>
-        /// Overridden by the nullable converter to prevent boxing of values by the JIT.
+        /// Performance optimization. Overridden by the nullable converter to prevent boxing of values by the JIT.
+        /// Although the JIT added support to detect this IL pattern (box+isinst+br), it still manifests in TryWrite().
         /// </summary>
-        internal virtual bool IsNull(in T value) => value == null;
+        internal virtual bool IsNull(in T value) => value is null;
 
+        /// <summary>
+        /// Performance optimization to determine if IsNull() above should be called.
+        /// </summary>
+        /// <returns></returns>
+        internal virtual bool IsNullableOfT() => false;
+
+#if NET6_0_OR_GREATER
+        [MethodImpl(MethodImplOptions.AggressiveOptimization)]
+#endif
         internal bool TryWrite(Utf8JsonWriter writer, in T value, JsonSerializerOptions options, ref WriteStack state)
         {
             if (writer.CurrentDepth >= options.EffectiveMaxDepth)
@@ -321,22 +352,38 @@ namespace System.Text.Json.Serialization
                 ThrowHelper.ThrowJsonException_SerializerCycleDetected(options.EffectiveMaxDepth);
             }
 
-            if (default(T) is null && !HandleNullOnWrite && IsNull(value))
+            // We do not pass null values to converters unless HandleNullOnWrite is true. Null values for properties were
+            // already handled in GetMemberAndWriteJson() so we don't need to check for IgnoreNullValues here.
+            if (_autoWriteNulls)
             {
-                // We do not pass null values to converters unless HandleNullOnWrite is true. Null values for properties were
-                // already handled in GetMemberAndWriteJson() so we don't need to check for IgnoreNullValues here.
-                writer.WriteNullValue();
-                return true;
+                // Optimized check to see if 'IsNull()' or 'is null' should be used.
+                if (_isNullableOfT)
+                {
+                    if (IsNull(value))
+                    {
+                        writer.WriteNullValue();
+                        return true;
+                    }
+                }
+                else if (value is null)
+                {
+                    writer.WriteNullValue();
+                    return true;
+                }
             }
 
             bool ignoreCyclesPopReference = false;
 
             if (
 #if NET5_0_OR_GREATER
-                !typeof(T).IsValueType && // treated as a constant by recent versions of the JIT.
+                // Short-circuit the check against "is not null"; treated as a constant by recent versions of the JIT.
+                !typeof(T).IsValueType &&
 #else
                 !IsValueType &&
 #endif
+                // Since we may have checked for a null value above we may have a redundant check here,
+                // but this seems to be better than trying to cache that value when considering all permutations:
+                // int?, int?(null value), int, object, object(null value)
                 value is not null)
             {
 
@@ -433,11 +480,12 @@ namespace System.Text.Json.Serialization
 
                 if (
 #if NET5_0_OR_GREATER
-                    !typeof(T).IsValueType && // treated as a constant by recent versions of the JIT.
+                    // Short-circuit the check against ignoreCyclesPopReference; treated as a constant by recent versions of the JIT.
+                    !typeof(T).IsValueType &&
 #endif
                     ignoreCyclesPopReference)
                 {
-                    // should only be entered if we're serializing instances
+                    // Should only be entered if we're serializing instances
                     // of type object using the internal object converter.
                     Debug.Assert(value?.GetType() == typeof(object) && IsInternalConverter);
                     state.ReferenceResolver.PopReferenceForCycleDetection();