--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.Node.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.Node.cs
@@ -149,7 +149,7 @@ namespace System.Text.Json
                 }
             }
 
-            return ReadSpan<TValue>(output.WrittenMemory.Span, jsonTypeInfo);
+            return ReadFromSpan<TValue>(output.WrittenMemory.Span, jsonTypeInfo);
         }
     }
 }