Primarily addresses serializer overhead of a given call, but there are general gains throughout.

Changes include:
- Manual inlinining \ lifting of serializer calls. This avoids some `if` statements and the smaller size appears to help the jitter.
- Avoids a cast against `JsonTypeInfo<T>` in unnecessary cases during serialization; this is expected to help perf for source-gen when falling back to the serializer.
- Use of `[MethodImplOptions.AggressiveOptimization]` in hot, looping methods.
- Optimized check for `null` when serializing either `Nullable<T>` and non-`Nullable<T>`.
- Misc others, including some consistency changes for variable and method names.

It primarily resolves this issue with overhead:
resolves https://github.com/dotnet/runtime/issues/56993
<details>
  <summary>Overhead</summary>

A class with no members:

5.0
|      Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------ |----------:|---------:|---------:|-------:|----------:|
|   Serialize |  85.53 ns | 0.375 ns | 0.351 ns |      - |         - |
| Deserialize | 268.87 ns | 1.627 ns | 1.522 ns | 0.0049 |      32 B |

6.0 before this PR
|      Method |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|------------ |---------:|--------:|--------:|-------:|----------:|
|   Serialize | 106.5 ns | 0.55 ns | 0.49 ns |      - |         - |
| Deserialize | 267.9 ns | 0.55 ns | 0.49 ns | 0.0049 |      32 B |

6.0 with this PR
|      Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------ |----------:|---------:|---------:|-------:|----------:|
|   Serialize |  85.88 ns | 0.527 ns | 0.493 ns |      - |         - |
| Deserialize | 263.78 ns | 2.824 ns | 2.503 ns | 0.0049 |      32 B |

Serialize overhead: 1.24x faster
Deserialize overhead: 1.015x faster

Also, if this test is changed to add 2 String properties:

5.0
|      Method |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|------------ |---------:|--------:|--------:|-------:|----------:|
|   Serialize | 196.3 ns | 1.31 ns | 1.16 ns |      - |         - |
| Deserialize | 341.1 ns | 0.74 ns | 0.70 ns | 0.0161 |     104 B |

6.0 before this PR
|      Method |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|------------ |---------:|--------:|--------:|-------:|----------:|
|   Serialize | 218.2 ns | 1.06 ns | 0.94 ns |      - |         - |
| Deserialize | 334.5 ns | 0.71 ns | 0.63 ns | 0.0161 |     104 B |

6.0 before this PR
|      Method |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|------------ |---------:|--------:|--------:|-------:|----------:|
|   Serialize | 198.4 ns | 2.19 ns | 2.05 ns |      - |         - |
| Deserialize | 324.2 ns | 2.35 ns | 2.08 ns | 0.0161 |     104 B |

Serialize: 1.1x faster compared to before this PR
Deserialize: 1.03x faster compared to before this PR
</details>

but also should resolve these (verification needed for some):

resolves https://github.com/dotnet/runtime/issues/57093
resolves https://github.com/dotnet/runtime/issues/52640
resolves https://github.com/dotnet/runtime/issues/52311
resolves https://github.com/dotnet/runtime/issues/43413

<details>
  <summary>Improvements in benchmarks with this PR</summary>

```
note: threshold is 2%
summary:
better: 31, geomean: 1.057
worse: 2, geomean: 1.071
total diff: 33
```

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromStream    |      1.11 |          1386.49 |          1540.57 | bimodal |
| System.Text.Json.Serialization.Tests.WriteJson<MyEventsListerViewModel>.Serializ |      1.03 |        377918.23 |        390056.88 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.WriteJson<Int32>.SerializeToUtf8Bytes       |      1.19 |           119.98 |           100.69 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Int32>.SerializeToString          |      1.16 |           137.06 |           117.90 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Int32>.SerializeObjectProperty    |      1.09 |           246.90 |           225.56 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeObjectPr |      1.09 |           505.15 |           462.27 |         |
| System.Text.Json.Serialization.Tests.WriteJson<ArrayList>.SerializeToString      |      1.08 |          7678.84 |          7140.75 |         |
| System.Text.Json.Serialization.Tests.WriteJson<SimpleStructWithProperties>.Seria |      1.07 |           251.04 |           234.99 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LargeStructWithProperties>.Serial |      1.06 |           622.98 |           586.76 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeToUtf8By |      1.06 |           303.54 |           286.05 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromUtf |      1.06 |           393.55 |           371.04 |         |
| System.Text.Json.Serialization.Tests.WriteJson<SimpleStructWithProperties>.Seria |      1.06 |           412.13 |           388.94 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LargeStructWithProperties>.Serial |      1.06 |           862.66 |           815.38 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Int32>.DeserializeFromString       |      1.06 |           124.69 |           118.10 |         |
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.05 |           559.86 |           530.84 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeToString |      1.05 |           331.31 |           314.80 |         |
| System.Text.Json.Serialization.Tests.WriteJson<SimpleStructWithProperties>.Seria |      1.05 |           225.08 |           214.83 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Location>.SerializeToString       |      1.05 |           810.09 |           774.23 |         |
| System.Text.Json.Serialization.Tests.WriteJson<ArrayList>.SerializeToUtf8Bytes   |      1.05 |          7249.32 |          6936.89 |         |
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.04 |           334.56 |           320.96 |         |
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeToUtf8Bytes  |      1.04 |           390.61 |           374.94 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromStr |      1.04 |           729.07 |           700.01 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromUtf8Bytes |      1.04 |          1076.71 |          1035.66 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Location>.SerializeToUtf8Bytes    |      1.04 |           779.50 |           750.04 |         |
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeToString     |      1.04 |           543.79 |           523.71 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LargeStructWithProperties>.Serial |      1.04 |           658.87 |           636.45 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Int32>.DeserializeFromUtf8Bytes    |      1.03 |            74.08 |            71.59 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LargeStructWithProperties>.Serial |      1.03 |           774.83 |           748.87 |         |
| System.Text.Json.Serialization.Tests.WriteJson<MyEventsListerViewModel>.Serializ |      1.03 |        379279.76 |        367081.10 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LargeStructWithProperties>.Deseria |      1.03 |          1145.78 |          1114.74 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LargeStructWithProperties>.Deseria |      1.03 |          1556.95 |          1515.01 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Int32>.DeserializeFromStream       |      1.03 |           298.83 |           290.81 |         |
| System.Text.Json.Serialization.Tests.WriteJson<HashSet<String>>.SerializeToStrin |      1.03 |          6444.12 |          6282.43 |         |
</details>


<details>
  <summary>6.0 status (with this PR) of benchmarks compared to 5.0 </summary>

```
note: threshold is 2%
summary:
better: 22, geomean: 1.268
worse: 2, geomean: 1.065
total diff: 24
```

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.10 |           482.68 |           530.03 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromStr |      1.03 |           668.86 |           690.92 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromString   |      2.13 |         47511.77 |         22281.48 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromStream   |      2.10 |         47456.53 |         22649.40 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromUtf8Byte |      2.07 |         46860.63 |         22584.86 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromString   |      1.88 |         67083.49 |         35639.60 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromUtf8Byte |      1.76 |         63251.21 |         35890.00 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Hashtable>.DeserializeFromStream   |      1.74 |         63594.48 |         36651.60 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableDictionary<String, String |      1.14 |         45314.11 |         39891.80 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableDictionary<String, String |      1.12 |         45394.56 |         40385.56 |         |
| System.Text.Json.Serialization.Tests.ReadJson<HashSet<String>>.DeserializeFromUt |      1.11 |         11901.75 |         10722.24 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Dictionary<String, String>>.Deseri |      1.11 |         21191.19 |         19168.64 |         |
| System.Text.Json.Serialization.Tests.ReadJson<HashSet<String>>.DeserializeFromSt |      1.10 |         12139.81 |         11039.48 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableSortedDictionary<String,  |      1.10 |         75880.24 |         69099.21 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableDictionary<String, String |      1.09 |         47138.25 |         43118.00 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromStr |      1.09 |         29213.97 |         26797.27 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Dictionary<String, String>>.Deseri |      1.08 |         20351.18 |         18857.99 |         |
| System.Text.Json.Serialization.Tests.ReadJson<IndexViewModel>.DeserializeFromUtf |      1.07 |         27666.04 |         25944.15 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Dictionary<String, String>>.Deseri |      1.07 |         22345.68 |         20956.26 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromUtf8Bytes |      1.06 |          1105.41 |          1039.88 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromString    |      1.06 |          1178.27 |          1113.83 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableSortedDictionary<String,  |      1.05 |         75538.87 |         72250.75 |         |
| System.Text.Json.Serialization.Tests.ReadJson<BinaryData>.DeserializeFromString  |      1.04 |           587.58 |           564.27 |         |
| System.Text.Json.Serialization.Tests.ReadJson<HashSet<String>>.DeserializeFromSt |      1.03 |         12883.77 |         12530.95 |         |

</summary>